% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-class.R
\name{is_prosody_score}
\alias{is_prosody_score}
\title{Test whether an object is a prosody_score}
\usage{
is_prosody_score(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a prosody_score
}
