% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{is_syn_media}
\alias{is_syn_media}
\title{Test whether an object is a syn_media container}
\usage{
is_syn_media(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a syn_media container
}
\seealso{
Other media functions: 
\code{\link{media_type}()}
}
\concept{media functions}
