% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{is_syn_palette}
\alias{is_syn_palette}
\title{Test whether an object is a syn_palette}
\usage{
is_syn_palette(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical scalar.
}
\description{
Test whether an object is a syn_palette
}
\seealso{
Other palette functions: 
\code{\link{syn_palette}()},
\code{\link{syn_theme}()}
}
\concept{palette functions}
