% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media_type}
\alias{media_type}
\alias{media_content}
\title{Accessors for syn_media content and type}
\usage{
media_type(x)

media_content(x)
}
\arguments{
\item{x}{A \code{syn_media} object.}
}
\value{
\code{media_type()} returns the type tag (\code{"text"}, \code{"image"}, \code{"audio"},
\code{"score"}); \code{media_content()} returns the wrapped object.
}
\description{
Accessors for syn_media content and type
}
\examples{
m <- as_syn_media("some text")
media_type(m)
}
\seealso{
Other media functions: 
\code{\link{is_syn_media}()}
}
\concept{media functions}
