% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synesthR-package.R
\docType{package}
\name{synesthR-package}
\alias{synesthR}
\alias{synesthR-package}
\title{synesthR: Generate Synchronized Images and Music from Literary Text}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extracts linguistic features from literary text in German, English, and French, including a sentence-level sentiment arc, lexical diversity, rhythm, and punctuation density, and maps them deterministically onto visual and musical parameters. From a single shared feature score it renders a generative still image and a synchronized audio waveform, so that the picture and the music are two views of the same analysis. Colour palettes and 'ggplot2' themes can be extracted not only from text but from images and audio through a common interface, and media objects (text, image, audio, score) are passed between functions through a shared interchange container. An optional assistive layer can interpret results using a locally running large language model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-heller/synesthR}
  \item \url{https://r-heller.github.io/synesthR/}
  \item Report bugs at \url{https://github.com/r-heller/synesthR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

}
\keyword{internal}
